Move count_issues_by_snapshot to ground_truth recipe, delete dead critic_dev_run_costs
count_issues_by_snapshot had no production callers — move it to the
ground_truth recipe where it fits alongside TP/FP querying.

critic_dev_run_costs was fully dead (zero callers) and redundant with
the run_analysis recipe which covers the same domain with simpler ORM queries.
