+++ props/db/BUILD.bazel
@@ -186,8 +186,8 @@ py_test(
         ":database",
         ":examples",
         ":models",
-        ":query_builders",
         "//props:conftest",
+        "//props/agents/critic_dev/recipes:ground_truth",
         "//props/core:splits",
         "//props/testing/fixtures:runs",
         "@pypi//pytest",
